relocate package to follow newest versions
