+++ src/main/java/team/creative/creativecore/common/gui/creator/GuiCreator.java
@@ -1,4 +1,4 @@
-package team.creative.creativecore.common.gui.handler;
+package team.creative.creativecore.common.gui.creator;
 
 import java.util.function.BiFunction;
