+++ src/main/java/team/creative/creativecore/common/gui/creator/ItemGuiCreator.java
@@ -1,4 +1,4 @@
-package team.creative.creativecore.common.gui.handler;
+package team.creative.creativecore.common.gui.creator;
 
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.player.Player;
